errors: move traceback in modelkit.core
