+++ tests/test_validate.py
@@ -3,13 +3,10 @@ from typing import Any, Dict, List
 import pydantic
 import pytest
 
+from modelkit.core.errors import ItemValidationException, ReturnValueValidationException
 from modelkit.core.model import Model
 from modelkit.core.settings import LibrarySettings
 from modelkit.utils.pydantic import construct_recursive
-from modelkit.utils.traceback import (
-    ItemValidationException,
-    ReturnValueValidationException,
-)
 
 
 @pytest.mark.parametrize(
